@@ -1,4 +1,4 @@
-# Copyright (C) DATADVANCE, 2010-2021
+# Copyright (C) DATADVANCE, 2011-2022
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -37,26 +37,35 @@ NOTE: Links based on which this functionality is implemented:
   https://gist.github.com/tricoder42/af3d0337c1b33d82c1b32d12bd0265ec
 """
 
-
 import asyncio
-import concurrent
 import dataclasses
 import functools
+import inspect
 import logging
+import time
 import traceback
-import types
 import weakref
-from typing import Any, Callable, Dict, List, Optional, Sequence
+from typing import (
+    Any,
+    AsyncIterator,
+    Callable,
+    Dict,
+    List,
+    Optional,
+    Sequence,
+    Union,
+    cast,
+)
 
 import asgiref.sync
+import channels.db
 import channels.generic.websocket as ch_websocket
-import django.core.serializers
-import django.db
 import graphql
 import graphql.error
-import graphql.execution.executors.asyncio
-import promise
-import rx
+import graphql.error.graphql_error
+import graphql.execution
+import graphql.pyutils
+import graphql.utilities
 
 from .scope_as_context import ScopeAsContext
 from .serializer import Serializer
@@ -66,6 +75,64 @@ LOG = logging.getLogger(__name__)
 
 # WebSocket subprotocol used for the GraphQL.
 GRAPHQL_WS_SUBPROTOCOL = "graphql-ws"
+
+
+async def _subscribe(
+    schema: graphql.GraphQLSchema,
+    document: graphql.DocumentNode,
+    middleware_manager: graphql.MiddlewareManager,
+    root_value: Any = None,
+    context_value: Any = None,
+    variable_values: Optional[Dict[str, Any]] = None,
+    operation_name: Optional[str] = None,
+    field_resolver: Optional[graphql.GraphQLFieldResolver] = None,
+    subscribe_field_resolver: Optional[graphql.GraphQLFieldResolver] = None,
+) -> Union[AsyncIterator[graphql.ExecutionResult], graphql.ExecutionResult]:
+    """Create a GraphQL subscription.
+
+    This is a copy of function from graphql-core library. It is extended
+    with middleware_manager.
+    """
+    result_or_stream = await graphql.create_source_event_stream(
+        schema,
+        document,
+        root_value,
+        context_value,
+        variable_values,
+        operation_name,
+        subscribe_field_resolver,
+    )
+    if isinstance(result_or_stream, graphql.ExecutionResult):
+        return result_or_stream
+
+    async def map_source_to_response(payload: Any) -> graphql.ExecutionResult:
+        """Map source to response.
+
+        For each payload yielded from a subscription, map it over the
+        normal GraphQL :func:`~graphql.execute` function, with
+        `payload` as the `root_value`. This implements the
+        "MapSourceToResponseEvent" algorithm described in the GraphQL
+        specification. The :func:`~graphql.execute` function provides
+        the "ExecuteSubscriptionEvent" algorithm, as it is nearly
+        identical to the "ExecuteQuery" algorithm, for which
+        :func:`~graphql.execute` is also used.
+        """
+        result = await graphql.execute(
+            schema,
+            document,
+            payload,
+            context_value,
+            variable_values,
+            operation_name,
+            field_resolver,
+            middleware=middleware_manager,
+        )  # type: ignore
+        if inspect.isawaitable(result):
+            return cast(graphql.ExecutionResult, await result)
+        return cast(graphql.ExecutionResult, result)
+
+    # Map every source value to a ExecutionResult value as described above.
+    return graphql.MapAsyncIterator(result_or_stream, map_source_to_response)
 
 
 class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
@@ -86,7 +153,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     schema = None
 
     # The interval to send keepalive messages to the clients (seconds).
-    send_keepalive_every = None
+    send_keepalive_every: Optional[int] = None
 
     # Set to `True` to process requests (i.e. GraphQL documents) from
     # a client in order of arrival, which is the same as sending order,
@@ -96,46 +163,52 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # before it finishes the current one. Note that requests from
     # different clients (within different WebSocket connections)
     # are still processed asynchronously. Useful for tests.
-    strict_ordering = False
+    strict_ordering: bool = False
 
     # When set to `True` the server will send an empty data message in
     # response to the subscription. This is needed to let client know
     # when the subscription activates, so he can be sure he doesn't miss
     # any notifications. Disabled by default, cause this is an extension
     # to the original protocol and the client must be tuned accordingly.
-    confirm_subscriptions = False
+    confirm_subscriptions: bool = False
 
     # The message sent to the client when subscription activation
     # confirmation is enabled.
-    subscription_confirmation_message = {"data": None, "errors": None}
+    subscription_confirmation_message: Dict[str, Any] = {"data": None, "errors": None}
 
-    # The maximum number of threads designated for GraphQL requests
-    # processing. `None` means that default value is used. Default
-    # value depends on the Python version, check its documentation:
-    # https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.ThreadPoolExecutor
-    max_worker_threads = None
+    # The interval given to every operation / resolver to execute in
+    # seconds. If an operation or an resolver took longer than that a
+    # warning message will be printed in the LOG.
+    # `None` value (as well as zero) means "do not print warnings".
+    warn_if_operation_took_longer_then: Optional[int] = 1
+    warn_if_resolver_took_longer_then: Optional[int] = 1
 
     # The size of the subscription notification queue. If there are more
     # notifications (for a single subscription) than the given number,
     # then an oldest notification is dropped and a warning is logged.
-    subscription_notification_queue_limit = 1024
+    subscription_notification_queue_limit: int = 1024
 
     # A prefix of the Channel group names used for the subscription
     # notifications. You may change this to avoid name clashes in the
     # ASGI backend, e.g. in the Redis.
-    group_name_prefix = "GRAPHQL_WS_SUBSCRIPTION"
+    group_name_prefix: str = "GRAPHQL_WS_SUBSCRIPTION"
 
     # GraphQL middleware.
     # Typically a list of functions (callables) like:
     # ```python
-    # def my_middleware(next_middleware, root, info, *args, **kwds):
-    #     return next_middleware(root, info, *args, **kwds)
+    # async def my_middleware(next_middleware, root, info, *args, **kwds):
+    #     return await next_middleware(root, info, *args, **kwds)
     # ```
     # For more information read:
     # https://docs.graphene-python.org/en/latest/execution/middleware/#middleware
     middleware: Sequence = []
 
-    async def on_connect(self, payload):
+    # A function to execute synchronous code. By default we rely on
+    # the `channels.db.database_sync_to_async` function and the thread
+    # pool it uses internally.
+    db_sync_to_async: asgiref.sync.SyncToAsync = channels.db.database_sync_to_async
+
+    async def on_connect(self, user, payload):
         """Client connection handler.
 
         Called after CONNECTION_INIT message from client. Overwrite and
@@ -143,26 +216,57 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         when it's necessary.
 
         Args:
+            user: A Django user returned by `get_user`.
             payload: Payload from CONNECTION_INIT message.
-
         """
+
+    async def get_user(self) -> Optional[Any]:
+        """Get Django user instance associated with current connection.
+
+        Returns:
+            A Django user or None
+        """
+        return None
+
+    async def on_operation_start(
+        self, scope: Dict[str, Any], operation_id, content
+    ) -> bool:
+        """Process business logic before operation processing will start.
+
+        For example this hook can check that user session is active (not
+        yet expired).
+
+        If you return `False` from this method it is your responsibility
+        to call `self.send_json` here.
+
+        Args:
+            scope: Dict with connection information.
+                    Example: {
+                        'type': 'websocket',
+                        'path': 'graphql/',
+                        'query_string': b'',
+                        'headers': [],
+                        'subprotocols': ['graphql-ws'],
+                        'path_remaining': '',
+                        'url_route': {'args': (), 'kwargs': {}},
+                        '_scope': {...},
+                        'user': {...},
+                        'operation_id': '726294ea34344c58b1a298bc16cc9fbe',
+                    }
+            operation_id: Operation id
+            content: Dict with json received through websocket.
+        Returns:
+            True - if operation is allowed to start.
+            False - if operation is forbidden.
+        """
+        del scope, operation_id, content
+        return True
 
     # ------------------------------------------------------------------- IMPLEMENTATION
 
     # Subscription implementation may shall return this to tell consumer
     # to suppress subscription notification.
     SKIP = object()
-
-    # Threadpool used to process requests.
-    _workers = concurrent.futures.ThreadPoolExecutor(
-        thread_name_prefix="GraphqlWs.", max_workers=max_worker_threads
-    )
-
-    # Disable promise "trampoline" (whatever it means), cause when it
-    # is enabled (which is by default) the promise is not thread-safe.
-    # This leads to hanging threads and unanswered requests.
-    # https://github.com/syrusakbary/promise/issues/57#issuecomment-406778476
-    promise.promise.async_instance.disable_trampoline()
 
     # Structure that holds subscription information.
     @dataclasses.dataclass
@@ -176,21 +280,24 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         # The subscription notification queue. Required to preserve the
         # order of notifications within a single subscription.
         notification_queue: asyncio.Queue
-        # A task which listens to the notification queue and notifies
-        # clients. Task triggers subscription `publish` resolvers.
-        notifier_task: asyncio.Task
         # The callback to invoke when client unsubscribes.
         unsubscribed_callback: Callable[[], None]
 
     def __init__(self, *args, **kwargs):
         """Consumer constructor."""
+
         assert self.schema is not None, (
             "An attribute 'schema' is not set! Subclasses must specify "
             "the schema which processes GraphQL subscription queries."
         )
 
         # Registry of active (subscribed) subscriptions.
-        self._subscriptions = {}  # {'<sid>': '<SubInf>', ...}
+        self._subscriptions: Dict[
+            int, GraphqlWsConsumer._SubInf
+        ] = {}  # {'<sid>': '<SubInf>', ...}
+        # A task which processes incoming messages and sends them to
+        # clients.
+        self._subscription_consumers: Dict[int, asyncio.Task] = {}
         self._sids_by_group = {}  # {'<grp>': ['<sid0>', '<sid1>', ...], ...}
 
         # Task that sends keepalive messages periodically.
@@ -210,6 +317,14 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         # requests. Since the collection is weak, it automatically
         # throws away items when locks are garbage collected.
         self._operation_locks = weakref.WeakValueDictionary()
+
+        # Offload the GraphQL processing to the worker thread cause
+        # according to our experiments even GraphQL document parsing
+        # (as well as validation) may take a while (and depends
+        # approx. linearly on the size of the selection set).
+        self._async_deserialize = self.db_sync_to_async(Serializer.deserialize)
+        self._async_graphql_parse = self.db_sync_to_async(graphql.parse)
+        self._async_graphql_validate = self.db_sync_to_async(graphql.validate)
 
         super().__init__(*args, **kwargs)
 
@@ -276,12 +391,22 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             self._keepalive_task.cancel()
             waitlist += [self._keepalive_task]
 
+        # Stop subscription event processing.
+        for consumer_task in self._subscription_consumers.values():
+            consumer_task.cancel()
+            waitlist.append(consumer_task)
+
+        # Wait for tasks to stop.
         if waitlist:
+            waitlist = [asyncio.ensure_future(task) for task in waitlist]
             await asyncio.wait(waitlist)
 
         self._subscriptions.clear()
         self._sids_by_group.clear()
         self._background_tasks.clear()
+        self._subscription_consumers.clear()
+        self._keepalive_task = None
+        self._operation_locks = weakref.WeakValueDictionary()
 
     async def receive_json(self, content):  # pylint: disable=arguments-differ
         """Process WebSocket message received from the client.
@@ -313,8 +438,13 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         elif msg_type == "START":
             op_id = content["id"]
+
+            allowed_to_start = await self.on_operation_start(self.scope, op_id, content)
+            if not allowed_to_start:
+                return
+
             # Create and lock a mutex for this particular operation id,
-            # so STOP processing for the same operation id will wail
+            # so STOP processing for the same operation id will wait
             # until START processing finishes. Locks are stored in a
             # weak collection so we do not have to manually clean it up.
             if op_id in self._operation_locks:
@@ -367,9 +497,12 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         Method is called when new `broadcast` message received from the
         Channels group. The message is sent by `Subscription.broadcast`.
-        Here we figure out the group message received from and trigger
-        the observable which makes the subscription process the query
-        and notify the client.
+
+        We search for target group. Then we take a list of subscriptions
+        registered for that group. And then we pushing the message to
+        subscriptions notification queues. The task that read
+        notifications from those queues will receive the message and
+        will send it to the client.
 
         NOTE: There is an issue in the `channels_redis` implementation
         which lead to the possibility to receive broadcast messages in
@@ -456,7 +589,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         # unsubscription.
         await asyncio.wait(
             [
-                self.receive_json({"type": "stop", "id": sid})
+                asyncio.ensure_future(self.receive_json({"type": "stop", "id": sid}))
                 for sid in self._sids_by_group[group]
             ]
         )
@@ -478,9 +611,10 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         try:
             # Notify subclass a new client is connected.
-            await self.on_connect(payload)
+            user = await self.get_user()
+            await self.on_connect(user, payload)
         except Exception as ex:  # pylint: disable=broad-except
-            LOG.error(str(ex))
+            LOG.exception(str(ex))
             await self._send_gql_connection_error(ex)
             # Close the connection.
             # NOTE: We use the 4000 code because there are two reasons:
@@ -556,86 +690,141 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             # useful for some use cases, take a loot at the issue from
             # more details:
             # https://github.com/datadvance/DjangoChannelsGraphqlWs/issues/41#issuecomment-663951595
-            context["channel_name"] = self.channel_name
+            context.channel_name = self.channel_name
+            context.db_sync_to_async = self.db_sync_to_async
+            context.group_name_prefix = self.group_name_prefix
 
-            # The `register` will be called from the worker thread
-            # (spawned for a GraphQL processing) when a client
-            # subscribes. So we "synchronize" it with `async_to_sync`.
-            # NOTE: It is important to invoke `async_to_sync` in the
-            # thread with the eventloop which serves this consumer. This
-            # assures that IO operations is performed within a single
-            # eventloop.
-            register_subscription = asgiref.sync.async_to_sync(
-                functools.partial(self._register_subscription, operation_id)
-            )
+            async def async_resolver_middleware(
+                next_middleware, root, info: graphql.GraphQLResolveInfo, *args, **kwds
+            ):
+                """Wraps resolver calls with `db_sync_to_async`.
 
-            def register_middleware(next_middleware, root, info, *args, **kwds):
-                """Transfers registration function to the `_subscribe`.
+                It is highly probable that resolvers will invoke
+                blocking DB operations and might be slow.
 
-                Check that `next_middleware` is precisely a subscription
-                handler `Subscription._subscribe` and hack it's `root`
-                to deliver the subscription registration function.
+                By wrapping resolvers with `db_sync_to_async` this
+                middleware ensures us that thouse slow operations will
+                be executed in separate worker thread.
+
+                Also this middleware able to track execution time of
+                resolvers and print warnings about slow resolvers.
                 """
+                if asyncio.iscoroutinefunction(next_middleware):
+                    next_func = next_middleware
+                else:
+                    next_func = self.db_sync_to_async(next_middleware)
 
-                # Avoid circular imports with local import.
-                # pylint: disable=import-outside-toplevel,cyclic-import
-                from .subscription import Subscription
+                if self.warn_if_resolver_took_longer_then:
+                    start_time = time.perf_counter()
 
-                # We do not expose `Subscription._subscribe` because
-                # `Subscription` is a public interface class. So it OK
-                # to get `Subscription._subscribe` here - we are tightly
-                # coupled anyway.
-                # pylint: disable=protected-access
-                if (
-                    getattr(next_middleware, "__func__", None)
-                    is Subscription._subscribe.__func__
-                ):
-                    root = types.SimpleNamespace(
-                        real_root=root, register_subscription=register_subscription
-                    )
-                return next_middleware(root, info, *args, **kwds)
+                result = next_func(root, info, *args, **kwds)
+                if inspect.isawaitable(result):
+                    result = await result
+
+                if self.warn_if_resolver_took_longer_then:
+                    duration = time.perf_counter() - start_time
+                    if duration >= self.warn_if_resolver_took_longer_then:
+                        # Trying to do some magic to show nice looking
+                        # function name in the log. Unfortunately this
+                        # is not always possible.
+                        try:
+                            if isinstance(next_middleware, functools.partial):
+                                next_middleware_nice_name = (
+                                    f"{next_middleware.func.__qualname__}"
+                                )
+                            elif hasattr(next_middleware, "__self__"):
+                                # it is a bound method with self variable?
+                                next_middleware_nice_name = (
+                                    f"{next_middleware.__self__.__qualname__}."
+                                    f"{next_middleware.__qualname__}"
+                                )
+                            else:
+                                next_middleware_nice_name = (
+                                    f"{next_middleware.__qualname__}"
+                                )
+                        except Exception:  # pylint: disable=broad-except
+                            LOG.exception(
+                                "Failed to form nice name for a function %s",
+                                next_middleware,
+                            )
+                            next_middleware_nice_name = str(next_middleware)
+
+                        parent_type = info.parent_type.name
+                        field_name = info.field_name
+
+                        LOG.warning(
+                            "Execution time of resolver %s "
+                            "for %s.%s is %.6f "
+                            "(operation_id=%s, op_name=%r)",
+                            next_middleware_nice_name,
+                            parent_type,
+                            field_name,
+                            duration,
+                            operation_id,
+                            op_name,
+                        )
+                return result
 
             # Process the request with Graphene/GraphQL.
 
-            # Offload the GraphQL processing to the worker thread cause
-            # according to our experiments even GraphQL document parsing
-            # may take a while (and depends approx. linearly on the size
-            # of the selection set). Moreover it is highly probable that
-            # resolvers will invoke blocking DB operations so it is
-            # better to offload the whole thing to a worker thread.
-            # NOTE: The `lambda` is required to force `AsyncioExecutor`
-            # take an eventloop from the worker thread, not the current
-            # one.
-            result = await self._run_in_worker(
-                lambda: graphql.graphql(
-                    self.schema,
-                    request_string=query,
-                    operation_name=op_name,
-                    variables=variables,
-                    context=context,
-                    # NOTE: Wrap with `wrap_in_promise=False`, otherwise
-                    # it raises `GraphQLError` with message:
-                    # "Subscription must return Async Iterable or
-                    # Observable. Received: <Promise...". I do not get
-                    # why it wraps it in promise by default.
-                    middleware=graphql.middlewares(
-                        register_middleware, *self.middleware, wrap_in_promise=False
-                    ),
-                    allow_subscriptions=True,
-                    executor=graphql.execution.executors.asyncio.AsyncioExecutor(),
-                )
+            # Do parsing
+            try:
+                document_ast = await self._async_graphql_parse(query)
+            except graphql.GraphQLError as error:
+                LOG.exception("Graphql parse failed")
+                raise error
+
+            # Do validation
+            validation_errors: List[
+                graphql.GraphQLError
+            ] = await self._async_graphql_validate(
+                self.schema.graphql_schema, document_ast
             )
+            if validation_errors:
+                LOG.warning(
+                    "Graphql validation failed with error: %s", validation_errors
+                )
+                await self._send_gql_data(operation_id, None, validation_errors)
+                await self._send_gql_complete(operation_id)
+                return
 
-        except Exception:  # pylint: disable=broad-except
-            # Something is wrong - send error message.
-            await self._send_gql_error(operation_id, traceback.format_exc())
+            op_ast = graphql.utilities.get_operation_ast(
+                document_ast, operation_name=op_name
+            )
+            # If the operation is subscription.
+            if op_ast.operation == graphql.language.ast.OperationType.SUBSCRIPTION:
+                context.operation_id = operation_id
+                # The `register_subscription` is called when a client
+                # subscribes.
+                context.register_subscription = functools.partial(
+                    self._register_subscription, operation_id
+                )
+                context.middleware_manager_for_subscriptions = (
+                    graphql.MiddlewareManager(
+                        async_resolver_middleware, *self.middleware
+                    )
+                )
+                LOG.debug(
+                    "Starting new subscription (operation_id=%s, op_name=%s)",
+                    operation_id,
+                    op_name,
+                )
 
-        else:
-            # Receiving an observer means the subscription has been
-            # processed. Otherwise it is just regular query or mutation.
-            if isinstance(result, rx.Observable):
-                # Client subscribed to a subscription so we subscribe to
-                # the observable returned.
+                # Normally the `_subscribe` call returns an
+                # AsyncGenerator (stream) witch yields a ExecutionResult
+                # for every notification. But in case of error it might
+                # return just an ExecutionResult.
+                result = await _subscribe(
+                    self.schema.graphql_schema,
+                    document_ast,
+                    operation_name=op_name,
+                    root_value=None,
+                    variable_values=variables,
+                    context_value=context,
+                    middleware_manager=graphql.MiddlewareManager(
+                        async_resolver_middleware, *self.middleware
+                    ),
+                )
 
                 # Respond with the subscription activation message if
                 # enabled in the consumer configuration.
@@ -650,28 +839,139 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                         errors=self.subscription_confirmation_message["errors"],
                     )
 
-                # Subscribe to the observable.
-                # NOTE: The function `on_next` is called from a
-                # threadpool which processes all GraphQL requests. So we
-                # wrap `_send_gql_data` into `async_to_sync` here, so
-                # eventually `_send_gql_data` will be invoked from the
-                # current eventloop.
-                send_gql_data = asgiref.sync.async_to_sync(self._send_gql_data)
-                result.subscribe(
-                    lambda r: send_gql_data(operation_id, r.data, r.errors)
+                # If the result is an AsyncGenerator (stream), then
+                # listen for incoming events (notifications).
+                if hasattr(result, "__aiter__"):
+                    consumer_init_lock = asyncio.Lock()
+                    await consumer_init_lock.acquire()
+
+                    async def _consume_stream():
+                        consumer_init_lock.release()
+                        try:
+                            async for item in result:
+                                try:
+                                    await self._send_gql_data(
+                                        operation_id, item.data, item.errors
+                                    )
+                                except asyncio.CancelledError:
+                                    break
+                        except Exception:  # pylint: disable=broad-except
+                            LOG.exception(
+                                "Subscription events consumer got exception "
+                                "(operation_id=%s).",
+                                operation_id,
+                            )
+                            await self._send_gql_error(
+                                operation_id, traceback.format_exc()
+                            )
+
+                    # We need to end this task when client drops
+                    # connection or unsubscribes, so lets store it.
+                    self._subscription_consumers[operation_id] = asyncio.ensure_future(
+                        _consume_stream()
+                    )
+
+                    # We must be sure here that the subscription
+                    # initialization is finished and the stream consumer
+                    # is active before we exit this function. Because in
+                    # the outer scope with have locking mechanism of
+                    # start and stop operations. And we want to say
+                    # "subscription operation is started" only when it
+                    # actually is.
+                    # This allows us to avoid the race condition between
+                    # simultaneous subscribe and unsubscribe calls.
+                    await consumer_init_lock.acquire()
+                    consumer_init_lock.release()
+            # If the operation is query or mutation.
+            else:
+                LOG.debug(
+                    "Starting new operation (operation_id=%s, op_name=%s)",
+                    operation_id,
+                    op_name,
                 )
 
-            else:
-                # Respond to a query or mutation immediately.
+                if self.warn_if_operation_took_longer_then:
+                    start_time = time.perf_counter()
 
-                # If Graphene complains that the request is invalid
-                # (e.g. query contains syntax error), then the `data`
-                # argument is `None` and the 'errors' argument contains
-                # all errors that occurred before or during execution.
-                # The `result` is an instance of the `ExecutionResult`.
-                await self._send_gql_data(operation_id, result.data, result.errors)
-                # Tell the client that the request processing is over.
+                result = await graphql.graphql(
+                    self.schema.graphql_schema,
+                    source=query,
+                    root_value=None,
+                    operation_name=op_name,
+                    variable_values=variables,
+                    context_value=context,
+                    middleware=graphql.execution.MiddlewareManager(
+                        async_resolver_middleware, *self.middleware
+                    ),
+                )
+                if self.warn_if_operation_took_longer_then:
+                    duration = time.perf_counter() - start_time
+                    if duration >= self.warn_if_operation_took_longer_then:
+                        LOG.warning(
+                            "Execution time of operation is %.6f "
+                            "(operation_id=%r, op_name=%r). "
+                            "Debug log contains a full query and variables.",
+                            duration,
+                            operation_id,
+                            op_name,
+                        )
+                        LOG.debug(
+                            "Execution time of operation is %.6f "
+                            "(operation_id=%r, op_name=%r, query=%r, vars=%r)",
+                            duration,
+                            operation_id,
+                            op_name,
+                            query,
+                            variables,
+                        )
+        except Exception as ex:  # pylint: disable=broad-except
+            # Something is wrong - send error message.
+            LOG.exception("Graphql operation error (operation_id=%s)", operation_id)
+
+            # In the version of the django-channels-graphql-ws 0.9.1 we
+            # returned syntax error in a specific way - as
+            # a `'type': 'data'`. Not a `'type': 'error'`. To provide
+            # backward compatability lets continue to return error in a
+            # same way.
+            #
+            # If Graphene complains that the request is invalid
+            # (e.g. query contains syntax error), then the `data`
+            # argument is `None` and the 'errors' argument contains
+            # all errors that occurred before or during execution.
+            if isinstance(ex, graphql.error.GraphQLSyntaxError):
+                # JSON for SyntaxError should look like this: {
+                #     'type': 'data',
+                #     'id': operation_id,
+                #     'payload': {
+                #          'data': None,
+                #          'errors': [{
+                #              'message': 'error message',
+                #              'locations': [{'line': 1, 'column': 1}]
+                #           }]
+                #     }
+                # }
+                await self._send_gql_data(operation_id, None, [ex])
                 await self._send_gql_complete(operation_id)
+            else:
+                await self._send_gql_error(operation_id, traceback.format_exc())
+        else:
+            # This is the "else" block of the "try-catch". Here we
+            # finalizing processing of the `result` value. And, if
+            # required, sending response to the client.
+
+            # Receiving an AsyncIterator (stream) means the subscription
+            # has been processed. We should not send anything to the
+            # client.
+            if hasattr(result, "__aiter__"):
+                return
+
+            # Respond to a query or mutation immediately.
+
+            # The `result` is an instance of the `ExecutionResult`.
+            await self._send_gql_data(operation_id, result.data, result.errors)
+
+            # Tell the client that the request processing is over.
+            await self._send_gql_complete(operation_id)
 
     async def _register_subscription(
         self,
@@ -684,67 +984,45 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         """Register a new subscription when client subscribes.
 
         This function is called (by the subscription implementation -
-        "resolver" function `_subscribe`) when a client subscribes to
-        register necessary callbacks.
+        the "resolver" function `_subscribe`) when a client subscribes
+        to register necessary callbacks.
 
         Args:
             operation_id: If of the protocol operation.
             groups: A list of subscription group names to put the
                 subscription into.
-            publish_callback: Called to process the notification (invoke
-                GraphQL resolvers). The callback will be associated with
-                a Channels groups used to deliver messages to the
-                subscription groups current subscription belongs to.
+            publish_callback: Called to process the notification (and to
+                invoke GraphQL resolvers). The callback will be
+                associated with a Channels groups used to deliver
+                messages to the subscription groups current subscription
+                belongs to.
             unsubscribed_callback: Called to notify when a client
                 unsubscribes.
-            notification_queue_limit: LImit for the subscribtion
+            notification_queue_limit: Limit for the subscription
                 notification queue. Default is used if not set.
-
         """
         # NOTE: It is important to invoke `group_add` from an
-        # eventloop which serves the current consumer. Besides
-        # this is a "proper" way of doing async things, it is
-        # also crucial for at least one Channels layer
-        # implementation - for the `channels_rabbitmq`. See:
+        # eventloop which serves the current consumer.  Besides this is
+        # a "proper" way of doing async things, it is also crucial for
+        # at least one Channels layer implementation - for the
+        # `channels_rabbitmq`. See:
         # https://github.com/CJWorkbench/channels_rabbitmq/issues/3
         # Moreover this allows us to access the `_subscriptions` and
         # `_sids_by_group` without any locks.
         self._assert_thread()
-
-        # The subject we will trigger on the `broadcast` message.
-        trigger = rx.subjects.Subject()
+        LOG.debug(
+            "Register new subscription (operation_id=%s, groups=%s)",
+            operation_id,
+            groups,
+        )
 
         # The subscription notification queue.
         queue_size = notification_queue_limit
         if not queue_size or queue_size < 0:
             queue_size = self.subscription_notification_queue_limit
+
+        # Note: asyncio.Queue is not thread-safe!
         notification_queue = asyncio.Queue(maxsize=queue_size)
-
-        # Start an endless task which listens the `notification_queue`
-        # and invokes subscription "resolver" on new notifications.
-        async def notifier():
-            """Watch the notification queue and notify clients."""
-
-            # Assert we run in a proper thread.
-            self._assert_thread()
-            while True:
-                payload = await notification_queue.get()
-                # Run a subscription's `publish` method (invoked by the
-                # `trigger.on_next` function) within the threadpool used
-                # for processing other GraphQL resolver functions.
-                # NOTE: `lambda` is important to run the deserialization
-                # in the worker thread as well.
-                await self._run_in_worker(
-                    lambda: trigger.on_next(Serializer.deserialize(payload))
-                )
-                # Message processed. This allows `Queue.join` to work.
-                notification_queue.task_done()
-
-        # Enqueue the `publish` method execution. But do not notify
-        # clients when `publish` returns `SKIP`.
-        stream = trigger.map(publish_callback).filter(  # pylint: disable=no-member
-            lambda publish_returned: publish_returned is not self.SKIP
-        )
 
         # Start listening for broadcasts (subscribe to the Channels
         # groups), spawn the notification processing task and put
@@ -754,19 +1032,30 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         waitlist = []
         for group in groups:
             self._sids_by_group.setdefault(group, []).append(operation_id)
-            waitlist.append(self._channel_layer.group_add(group, self.channel_name))
-        notifier_task = self._spawn_background_task(notifier())
+            waitlist.append(
+                asyncio.create_task(
+                    self._channel_layer.group_add(group, self.channel_name)
+                )
+            )
         self._subscriptions[operation_id] = self._SubInf(
             groups=groups,
             sid=operation_id,
             unsubscribed_callback=unsubscribed_callback,
             notification_queue=notification_queue,
-            notifier_task=notifier_task,
         )
 
         await asyncio.wait(waitlist)
 
-        return stream
+        # For each notification (event) yielded from this function
+        # the "_subscribe" function will call normal `graphql.execute`.
+        # Middlewares will be executed for that call.
+        while True:
+            payload = await notification_queue.get()
+            data = await self._async_deserialize(payload)
+            ret = await publish_callback(data)
+            if ret != self.SKIP:
+                yield ret
+            notification_queue.task_done()
 
     async def _on_gql_stop(self, operation_id):
         """Process the STOP message.
@@ -780,6 +1069,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         # Assert we run in a proper thread. In particular, we can access
         # the `_subscriptions` and `_sids_by_group` without any locks.
         self._assert_thread()
+        LOG.debug("Stop handling operation %r (unsubscribe)", operation_id)
 
         # Currently only subscriptions can be stopped. But we see but
         # some clients (e.g. GraphiQL) send the stop message even for
@@ -788,14 +1078,18 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         if operation_id not in self._subscriptions:
             return
 
-        # Unsubscribe:
-        # - Throw away the subscription from the registry.
-        # - Cancel the task which watches over the notification queue.
-        # - Stop listening corresponding groups.
         waitlist = []
+
+        # Remove the subscription from the registry.
         subinf = self._subscriptions.pop(operation_id)
-        subinf.notifier_task.cancel()
-        waitlist.append(subinf.notifier_task)
+
+        # Cancel the task which watches the notification queue.
+        consumer_task = self._subscription_consumers.pop(operation_id, None)
+        if consumer_task:
+            consumer_task.cancel()
+            waitlist.append(consumer_task)
+
+        # Stop listening for corresponding groups.
         for group in subinf.groups:
             # Remove the subscription from groups it belongs to. Remove
             # the group itself from the `_sids_by_group` if there are no
@@ -811,14 +1105,15 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
                 waitlist.append(
                     self._channel_layer.group_discard(group, self.channel_name)
                 )
+
+        waitlist = [asyncio.ensure_future(task) for task in waitlist]
         await asyncio.wait(waitlist)
 
-        # Call the subscription class `unsubscribed` handler in a worker
-        # thread, cause it may invoke long-running synchronous tasks.
-        await self._run_in_worker(subinf.unsubscribed_callback)
+        await subinf.unsubscribed_callback()
 
         # Send the unsubscription confirmation message.
         await self._send_gql_complete(operation_id)
+        LOG.debug("Operation %s stop - done", operation_id)
 
     # -------------------------------------------------------- GRAPHQL PROTOCOL MESSAGES
 
@@ -835,7 +1130,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
         )
 
     async def _send_gql_data(
-        self, operation_id, data: dict, errors: Optional[Sequence[Exception]]
+        self, operation_id, data: Optional[Dict], errors: Optional[Sequence[Exception]]
     ):
         """Send GraphQL `data` message to the client.
 
@@ -843,7 +1138,6 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             data: Dict with GraphQL query response.
             errors: List of exceptions occurred during processing the
                 GraphQL query. (Errors happened in resolvers.)
-
         """
         self._assert_thread()
         # Log errors with tracebacks so we can understand what happened
@@ -852,14 +1146,8 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
             # Typical exception here is `GraphQLLocatedError` which has
             # reference to the original error raised from a resolver.
             tb = ex.__traceback__
-            if (
-                isinstance(ex, graphql.error.located_error.GraphQLLocatedError)
-                and ex.original_error is not None
-            ):
-                tb = ex.stack
-                ex = ex.original_error
-            LOG.error(
-                "GraphQL resolver failed on operation with id=%s:\n%s",
+            LOG.warning(
+                "GraphQL resolver failed (operation_id=%s):\n%s",
                 operation_id,
                 "".join(traceback.format_exception(type(ex), ex, tb)).strip(),
             )
@@ -898,7 +1186,7 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
 
         """
         self._assert_thread()
-        LOG.error("GraphQL query processing error: %s", error)
+        LOG.warning("Send error (operation_id=%s): %s", operation_id, error)
         await self.send_json(
             {"type": "error", "id": operation_id, "payload": {"errors": [error]}}
         )
@@ -921,63 +1209,15 @@ class GraphqlWsConsumer(ch_websocket.AsyncJsonWebsocketConsumer):
     # ------------------------------------------------------------------------ AUXILIARY
 
     @staticmethod
-    def _format_error(error: Exception) -> Dict[str, Any]:
+    def _format_error(error: Exception) -> graphql.GraphQLFormattedError:
         """Format given exception `error` to send over a network."""
         if isinstance(error, graphql.error.GraphQLError):
-            return dict(graphql.error.format_error(error))
+            return error.formatted
 
-        return {"message": f"{type(error).__name__}: {str(error)}"}
-
-    async def _run_in_worker(self, func):
-        """Run a function in a thread with an event loop.
-
-        Run the `func` in a thread within a thread pool and wait for it
-        to finish. Each such thread has initialized event loop which is
-        NOT running. So the `func` can use it, for example by running
-        `asyncio.get_event_loop().run_until_complete(...)`. We also
-        cleanup Django database connections when `func` finishes.
-
-        Args:
-            func: The function to run in a thread.
-
-        Returns:
-            The result of the `func` invocation.
-
-        """
-        # Assert we run in a proper thread.
-        self._assert_thread()
-
-        def thread_func():
-            """Wrap the `func` to init eventloop and to cleanup."""
-            # Create an eventloop if this worker thread does not have
-            # one yet. Eventually each worker will have its own
-            # eventloop so `func` can use it.
-            try:
-                loop = asyncio.get_event_loop()
-            except RuntimeError:
-                loop = asyncio.new_event_loop()
-                # Force the event loop of the worker thread have a
-                # single-threaded default executor, otherwise the total
-                # number of threads may (and will) grow up to the value
-                # of `default_max_workers * default_max_workers` for
-                # each `GraphqlWsConsumer` subclass!
-                loop.set_default_executor(
-                    concurrent.futures.ThreadPoolExecutor(max_workers=1)
-                )
-                asyncio.set_event_loop(loop)
-
-            # Run a given function in a thread.
-            try:
-                return func()
-            finally:
-                # The `func` can open the connections to the database
-                # so let's close them. Django Channels does the same in
-                # the `channels.db.database_sync_to_async`.
-                django.db.close_old_connections()
-
-        return await asyncio.get_event_loop().run_in_executor(
-            self._workers, thread_func
-        )
+        return {
+            "message": f"{type(error).__name__}: {str(error)}",
+            "extensions": {"code": type(error).__name__},
+        }
 
     def _assert_thread(self):
         """Assert called from our thread with the eventloop."""
