@@ -1,4 +1,4 @@
-# Copyright (C) DATADVANCE, 2010-2021
+# Copyright (C) DATADVANCE, 2010-2022
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -196,6 +196,10 @@ async def test_custom_confirmation_message(gql):
 
 
 # ------------------------------------------------------------------- TEST GRAPHQL SETUP
+
+
+class Query(graphene.ObjectType):
+    """Root GraphQL query."""
 
 
 class Trigger(graphene.Mutation):
